requirements.txt
fastapi
httpx
sqlalchemy
mysql-connector-python
pydantic
uvicorn
alembic
pymysql
PyJWT